chore: Migrate to TypeScript
