+++ .gitignore
@@ -60,6 +60,6 @@ typings/
 # next.js build output
 .next
 
-index.js
+dist/
 !babel.config.js
 package-lock.json
