--- babel.config.js
@@ -1,15 +1,0 @@
-module.exports = (api) => {
-  api.cache(true)
-  return {
-    presets: [
-      [
-        '@babel/preset-env',
-        {
-          targets: {
-            node: '10'
-          }
-        }
-      ]
-    ]
-  }
-}
