+++ jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  verbose: true,
+  testTimeout: 10000,
+  transform: {
+    '^.+\\.ts$': 'ts-jest'
+  }
+}
