+++ src/index.ts
@@ -1,0 +1,3 @@
+import { plugin } from './plugin'
+
+export default plugin
